% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kms_operations.R
\name{cancel_key_deletion}
\alias{cancel_key_deletion}
\title{Cancels the deletion of a customer master key (CMK)}
\usage{
cancel_key_deletion(KeyId)
}
\arguments{
\item{KeyId}{[required] The unique identifier for the customer master key (CMK) for which to cancel deletion.

Specify the key ID or the Amazon Resource Name (ARN) of the CMK.

For example:
\itemize{
\item Key ID: \code{1234abcd-12ab-34cd-56ef-1234567890ab}
\item Key ARN: \code{arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab}
}

To get the key ID and key ARN for a CMK, use ListKeys or DescribeKey.}
}
\description{
Cancels the deletion of a customer master key (CMK). When this operation is successful, the CMK is set to the \code{Disabled} state. To enable a CMK, use EnableKey. You cannot perform this operation on a CMK in a different AWS account.
}
\details{
For more information about scheduling and canceling deletion of a CMK, see \href{http://docs.aws.amazon.com/kms/latest/developerguide/deleting-keys.html}{Deleting Customer Master Keys} in the \emph{AWS Key Management Service Developer Guide}.

The result of this operation varies with the key state of the CMK. For details, see \href{http://docs.aws.amazon.com/kms/latest/developerguide/key-state.html}{How Key State Affects Use of a Customer Master Key} in the \emph{AWS Key Management Service Developer Guide}.
}
\section{Accepted Parameters}{
\preformatted{cancel_key_deletion(
  KeyId = "string"
)
}
}

\examples{
# The following example cancels deletion of the specified CMK.
\donttest{cancel_key_deletion(
  KeyId = "1234abcd-12ab-34cd-56ef-1234567890ab"
)}

}
