% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kms_operations.R
\name{describe_custom_key_stores}
\alias{describe_custom_key_stores}
\title{Gets information about \link{custom key stores}(http://docs}
\usage{
describe_custom_key_stores(
  CustomKeyStoreId = "string",
  CustomKeyStoreName = "string",
  Limit = 123,
  Marker = "string"
)
}
\arguments{
\item{CustomKeyStoreId}{Gets only information about the specified custom key store. Enter the key store ID.

By default, this operation gets information about all custom key stores in the account and region. To limit the output to a particular custom key store, you can use either the \code{CustomKeyStoreId} or \code{CustomKeyStoreName} parameter, but not both.}

\item{CustomKeyStoreName}{Gets only information about the specified custom key store. Enter the friendly name of the custom key store.

By default, this operation gets information about all custom key stores in the account and region. To limit the output to a particular custom key store, you can use either the \code{CustomKeyStoreId} or \code{CustomKeyStoreName} parameter, but not both.}

\item{Limit}{Use this parameter to specify the maximum number of items to return. When this value is present, AWS KMS does not return more than the specified number of items, but it might return fewer.}

\item{Marker}{Use this parameter in a subsequent request after you receive a response with truncated results. Set it to the value of \code{NextMarker} from the truncated response you just received.}
}
\description{
Gets information about \href{http://docs.aws.amazon.com/kms/latest/developerguide/key-store-overview.html}{custom key stores} in the account and region.
}
\details{
This operation is part of the \href{http://docs.aws.amazon.com/kms/latest/developerguide/custom-key-store-overview.html}{Custom Key Store feature} feature in AWS KMS, which combines the convenience and extensive integration of AWS KMS with the isolation and control of a single-tenant key store.

By default, this operation returns information about all custom key stores in the account and region. To get only information about a particular custom key store, use either the \code{CustomKeyStoreName} or \code{CustomKeyStoreId} parameter (but not both).

To determine whether the custom key store is connected to its AWS CloudHSM cluster, use the \code{ConnectionState} element in the response. If an attempt to connect the custom key store failed, the \code{ConnectionState} value is \code{FAILED} and the \code{ConnectionErrorCode} element in the response indicates the cause of the failure. For help interpreting the \code{ConnectionErrorCode}, see CustomKeyStoresListEntry.

Custom key stores have a \code{DISCONNECTED} connection state if the key store has never been connected or you use the DisconnectCustomKeyStore operation to disconnect it. If your custom key store state is \code{CONNECTED} but you are having trouble using it, make sure that its associated AWS CloudHSM cluster is active and contains the minimum number of HSMs required for the operation, if any.

For help repairing your custom key store, see the \href{http://docs.aws.amazon.com/kms/latest/developerguide/fix-keystore-html}{Troubleshooting Custom Key Stores} topic in the \emph{AWS Key Management Service Developer Guide}.
}
\examples{

}
