% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kms_operations.R
\name{disable_key_rotation}
\alias{disable_key_rotation}
\title{Disables \link{automatic rotation of the key material}(http://docs}
\usage{
disable_key_rotation(
  KeyId = "string"
)
}
\arguments{
\item{KeyId}{[required] A unique identifier for the customer master key (CMK).

Specify the key ID or the Amazon Resource Name (ARN) of the CMK.

For example:
\itemize{
\item Key ID: \code{1234abcd-12ab-34cd-56ef-1234567890ab}
\item Key ARN: \code{arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab}
}

To get the key ID and key ARN for a CMK, use ListKeys or DescribeKey.}
}
\description{
Disables \href{http://docs.aws.amazon.com/kms/latest/developerguide/rotate-keys.html}{automatic rotation of the key material} for the specified customer master key (CMK). You cannot perform this operation on a CMK in a different AWS account.
}
\details{
The result of this operation varies with the key state of the CMK. For details, see \href{http://docs.aws.amazon.com/kms/latest/developerguide/key-state.html}{How Key State Affects Use of a Customer Master Key} in the \emph{AWS Key Management Service Developer Guide}.
}
\examples{
# The following example disables automatic annual rotation of the key
# material for the specified CMK.
disable_key_rotation(
  KeyId = "1234abcd-12ab-34cd-56ef-1234567890ab"
)

}
