% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kms_operations.R
\name{generate_random}
\alias{generate_random}
\title{Returns a random byte string that is cryptographically secure}
\usage{
generate_random(NumberOfBytes = NULL)
}
\arguments{
\item{NumberOfBytes}{The length of the byte string.}
}
\description{
Returns a random byte string that is cryptographically secure.
}
\details{
For more information about entropy and random number generation, see the \href{https://d0.awsstatic.com/whitepapers/KMS-Cryptographic-Details.pdf}{AWS Key Management Service Cryptographic Details} whitepaper.
}
\examples{
# The following example uses AWS KMS to generate 32 bytes of random data.
generate_random(
  NumberOfBytes = 32L
)

}
