% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kms_package.R
\docType{package}
\name{paws.kms-package}
\alias{paws.kms}
\alias{paws.kms-package}
\title{paws.kms: AWS Key Management Service}
\description{
AWS Key Management Service (AWS KMS) is an
    encryption and key management web service. This guide describes the
    AWS KMS operations that you can call programmatically. For general
    information about AWS KMS, see the [AWS Key Management Service
    Developer
    Guide](http://docs.aws.amazon.com/kms/latest/developerguide/).
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
