% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kms_operations.R
\name{update_key_description}
\alias{update_key_description}
\title{Updates the description of a customer master key (CMK)}
\usage{
update_key_description(KeyId, Description)
}
\arguments{
\item{KeyId}{A unique identifier for the customer master key (CMK).

Specify the key ID or the Amazon Resource Name (ARN) of the CMK.

For example:
\itemize{
\item Key ID: \code{1234abcd-12ab-34cd-56ef-1234567890ab}
\item Key ARN: \code{arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab}
}

To get the key ID and key ARN for a CMK, use ListKeys or DescribeKey.}

\item{Description}{New description for the CMK.}
}
\description{
Updates the description of a customer master key (CMK). To see the description of a CMK, use DescribeKey.
}
\details{
You cannot perform this operation on a CMK in a different AWS account.

The result of this operation varies with the key state of the CMK. For details, see \href{http://docs.aws.amazon.com/kms/latest/developerguide/key-state.html}{How Key State Affects Use of a Customer Master Key} in the \emph{AWS Key Management Service Developer Guide}.
}
\examples{
# The following example updates the description of the specified CMK.
update_key_description(
  Description = "Example description that indicates the intended use of this CMK.",
  KeyId = "1234abcd-12ab-34cd-56ef-1234567890ab"
)

}
