% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kms_operations.R
\name{create_key}
\alias{create_key}
\title{Creates a customer master key (CMK) in the caller's AWS account}
\usage{
create_key(Policy = NULL, Description = NULL, KeyUsage = NULL,
  Origin = NULL, BypassPolicyLockoutSafetyCheck = NULL, Tags = NULL)
}
\arguments{
\item{Policy}{The key policy to attach to the CMK.

If you provide a key policy, it must meet the following criteria:
\itemize{
\item If you don't set \code{BypassPolicyLockoutSafetyCheck} to true, the key policy must allow the principal that is making the \code{CreateKey} request to make a subsequent PutKeyPolicy request on the CMK. This reduces the risk that the CMK becomes unmanageable. For more information, refer to the scenario in the \href{http://docs.aws.amazon.com/kms/latest/developerguide/key-policies.html#key-policy-default-allow-root-enable-iam}{Default Key Policy} section of the \emph{AWS Key Management Service Developer Guide}.
\item Each statement in the key policy must contain one or more principals. The principals in the key policy must exist and be visible to AWS KMS. When you create a new AWS principal (for example, an IAM user or role), you might need to enforce a delay before including the new principal in a key policy. The reason for this is that the new principal might not be immediately visible to AWS KMS. For more information, see \href{http://docs.aws.amazon.com/IAM/latest/UserGuide/troubleshoot_general.html#troubleshoot_general_eventual-consistency}{Changes that I make are not always immediately visible} in the \emph{AWS Identity and Access Management User Guide}.
}

If you do not provide a key policy, AWS KMS attaches a default key policy to the CMK. For more information, see \href{http://docs.aws.amazon.com/kms/latest/developerguide/key-policies.html#key-policy-default}{Default Key Policy} in the \emph{AWS Key Management Service Developer Guide}.

The key policy size limit is 32 kilobytes (32768 bytes).}

\item{Description}{A description of the CMK.

Use a description that helps you decide whether the CMK is appropriate for a task.}

\item{KeyUsage}{The intended use of the CMK.

You can use CMKs only for symmetric encryption and decryption.}

\item{Origin}{The source of the CMK's key material.

The default is \code{AWS_KMS}, which means AWS KMS creates the key material. When this parameter is set to \code{EXTERNAL}, the request creates a CMK without key material so that you can import key material from your existing key management infrastructure. For more information about importing key material into AWS KMS, see \href{http://docs.aws.amazon.com/kms/latest/developerguide/importing-keys.html}{Importing Key Material} in the \emph{AWS Key Management Service Developer Guide}.

The CMK's \code{Origin} is immutable and is set when the CMK is created.}

\item{BypassPolicyLockoutSafetyCheck}{A flag to indicate whether to bypass the key policy lockout safety check.

Setting this value to true increases the risk that the CMK becomes unmanageable. Do not set this value to true indiscriminately.

For more information, refer to the scenario in the \href{http://docs.aws.amazon.com/kms/latest/developerguide/key-policies.html#key-policy-default-allow-root-enable-iam}{Default Key Policy} section in the \emph{AWS Key Management Service Developer Guide}.

Use this parameter only when you include a policy in the request and you intend to prevent the principal that is making the request from making a subsequent PutKeyPolicy request on the CMK.

The default value is false.}

\item{Tags}{One or more tags. Each tag consists of a tag key and a tag value. Tag keys and tag values are both required, but tag values can be empty (null) strings.

Use this parameter to tag the CMK when it is created. Alternately, you can omit this parameter and instead tag the CMK after it is created using TagResource.}
}
\description{
Creates a customer master key (CMK) in the caller's AWS account.
}
\details{
You can use a CMK to encrypt small amounts of data (4 KiB or less) directly. But CMKs are more commonly used to encrypt data encryption keys (DEKs), which are used to encrypt raw data. For more information about DEKs and the difference between CMKs and DEKs, see the following:
\itemize{
\item The GenerateDataKey operation
\item \href{http://docs.aws.amazon.com/kms/latest/developerguide/concepts.html}{AWS Key Management Service Concepts} in the \emph{AWS Key Management Service Developer Guide}
}

You cannot use this operation to create a CMK in a different AWS account.
}
\section{Accepted Parameters}{
\preformatted{create_key(
  Policy = "string",
  Description = "string",
  KeyUsage = "ENCRYPT_DECRYPT",
  Origin = "AWS_KMS"|"EXTERNAL",
  BypassPolicyLockoutSafetyCheck = TRUE|FALSE,
  Tags = list(
    list(
      TagKey = "string",
      TagValue = "string"
    )
  )
)
}
}

\examples{
# The following example creates a CMK.
\donttest{create_key(
  Tags = list(
    list(
      TagKey = "CreatedBy",
      TagValue = "ExampleUser"
    )
  )
)}

}
